requirements.txt
# ============================================================
# Зависимости для Telegram Analytics Platform
# Монолитное приложение с модульной архитектурой
# ============================================================

# Telegram Parser Module (services/telegram_parser.py)
telethon>=1.34.0

# Report Generator Module (services/report_generator.py)
google-genai>=1.21.1
docxtpl==0.16.8
pytils==0.4.1

# LLM Providers (services/llm_providers.py)
openai>=1.0.0  # Для Polza.ai (OpenAI-compatible API)

# FastAPI Application (app.py, routers/)
fastapi==0.111.0
uvicorn==0.24.0

# Telegram Bot (telegram_bot.py)
aiogram==3.13.1
aiohttp==3.10.5

# Общие зависимости
requests>=2.31.0
psutil==5.9.8  # Мониторинг памяти для memory_manager
aiofiles>=23.0.0  # Асинхронное чтение/запись файлов

# Тестирование (dev dependencies)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # Для тестирования FastAPI
